+++ CHANGELOG.md
@@ -13,6 +13,9 @@ All notable changes to this project will be documented in this file.
 - [#2119] Bulk imported assets can also match filenames based on assigned classname, not just character id prefix
 - Debugger shows (logs) unhandled exceptions
 - [#2129] MEMORY and STACK_SIZE parameters now can be set via external variables FFDEC_MEMORY, FFDEC_STACK_SIZE
+- Saving Harman encrypted SWFs
+- Editing encrypted flag on header panel
+- `-encrypt` command on CLI for Harman encryption
 
 ### Fixed
 - [#2021], [#2000] Caret position in editors when using tabs and / or unicode